+++ core/repository-mock/src/main/java/app/kaito_dogi/mybrary/core/repository/mock/di/AuthRepositoryModule.kt
@@ -1,6 +1,7 @@
-package app.kaito_dogi.mybrary.core.repository
+package app.kaito_dogi.mybrary.core.repository.mock.di
 
 import app.kaito_dogi.mybrary.core.domain.repository.AuthRepository
+import app.kaito_dogi.mybrary.core.repository.mock.MockAuthRepository
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -12,5 +13,5 @@ import javax.inject.Singleton
 internal interface AuthRepositoryModule {
   @Singleton
   @Binds
-  fun bindAuthRepository(impl: DefaultAuthRepository): AuthRepository
+  fun bindAuthRepository(impl: MockAuthRepository): AuthRepository
 }
