+++ core/repository-mock/src/main/java/app/kaito_dogi/mybrary/core/repository/mock/di/BookRepositoryModule.kt
@@ -1,6 +1,7 @@
-package app.kaito_dogi.mybrary.core.repository
+package app.kaito_dogi.mybrary.core.repository.mock.di
 
 import app.kaito_dogi.mybrary.core.domain.repository.BookRepository
+import app.kaito_dogi.mybrary.core.repository.mock.MockBookRepository
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -12,5 +13,5 @@ import javax.inject.Singleton
 internal interface BookRepositoryModule {
   @Singleton
   @Binds
-  fun bindBookRepository(impl: DefaultBookRepository): BookRepository
+  fun bindBookRepository(mock: MockBookRepository): BookRepository
 }
