+++ core/repository-mock/src/main/java/app/kaito_dogi/mybrary/core/repository/mock/di/MyBookRepositoryModule.kt
@@ -1,6 +1,7 @@
-package app.kaito_dogi.mybrary.core.repository
+package app.kaito_dogi.mybrary.core.repository.mock.di
 
 import app.kaito_dogi.mybrary.core.domain.repository.MyBookRepository
+import app.kaito_dogi.mybrary.core.repository.mock.MockMyBookRepository
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -12,5 +13,5 @@ import javax.inject.Singleton
 internal interface MyBookRepositoryModule {
   @Singleton
   @Binds
-  fun bindMyBookRepository(impl: DefaultMyBookRepository): MyBookRepository
+  fun bindMyBookRepository(impl: MockMyBookRepository): MyBookRepository
 }
